pyproject.toml
[tool.poetry]
name = "jqboard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
lxml = "^4.9.1"
pywin32 = {version = "^304", platform="windows"}
Pillow = "^9.2.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
